% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_media.R
\name{add_new_social_media}
\alias{add_new_social_media}
\title{Create new social media template files, so a developer can extend episomer and support a new social media.
this function has to be called within episomer subfolder containing thr R subfolder after executing devtools::load_all()}
\usage{
add_new_social_media(social_media, social_media_ref = "bluesky")
}
\arguments{
\item{social_media}{Name of the new social media}

\item{social_media_ref}{Name of the reference social media}
}
\description{
Create new social media template files, so a developer can extend episomer and support a new social media.
this function has to be called within episomer subfolder containing thr R subfolder after executing devtools::load_all()
}
