% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R
\name{bluesky_check_token_validity}
\alias{bluesky_check_token_validity}
\title{check the current bluesky token validity}
\usage{
bluesky_check_token_validity(
  access_jwt,
  search_url = "https://bsky.social/xrpc/app.bsky.feed.searchPosts"
)
}
\arguments{
\item{access_jwt}{char, the jwt token provided by the bluesky api}

\item{search_url}{char, the url to test if the current token is still valid}
}
\description{
check the current bluesky token validity
}
