% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R
\name{bluesky_create_session}
\alias{bluesky_create_session}
\title{Get bearer token}
\usage{
bluesky_create_session(handle = NULL, password = NULL)
}
\arguments{
\item{handle}{bluesky handle}

\item{password}{bluesky password}
}
\value{
List with access_jwt and did
}
\description{
Get bearer token
}
\details{
You can find more information about the bluesky API here: \url{https://docs.bsky.app/docs/api/com-atproto-server-create-session}
}
\examples{
\dontrun{
bluesky_create_session()
}

}
