% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R
\name{bluesky_format_date}
\alias{bluesky_format_date}
\title{formats a date in the expected bluesky format}
\usage{
bluesky_format_date(datetime)
}
\arguments{
\item{datetime, }{the datetime to format}
}
\description{
formats a date in the expected bluesky format
}
