% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R
\name{bluesky_parse_date}
\alias{bluesky_parse_date}
\title{parse a date as provided by bluesky into a date_time object}
\usage{
bluesky_parse_date(date_input)
}
\arguments{
\item{date_input}{char, the text of a date}
}
\description{
parse a date as provided by bluesky into a date_time object
}
