% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R
\name{bluesky_parse_features}
\alias{bluesky_parse_features}
\title{extracts features from the post structure returned by the bluesky api}
\usage{
bluesky_parse_features(post)
}
\arguments{
\item{post}{list, the post to extract features from in bluesky format}
}
\description{
extracts features from the post structure returned by the bluesky api
}
