% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R
\name{bluesky_parse_quoted}
\alias{bluesky_parse_quoted}
\title{extracts the quotes information from a post in one of the many possible locations found in bluesky format}
\usage{
bluesky_parse_quoted(post)
}
\arguments{
\item{post}{list, the post to to extract data from in bluesky format}
}
\description{
extracts the quotes information from a post in one of the many possible locations found in bluesky format
}
