% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R
\name{bluesky_parse_response}
\alias{bluesky_parse_response}
\title{parses the response of bluesky API into the standardised episomer format. For more details see the vignette.}
\usage{
bluesky_parse_response(response, current_min_date)
}
\arguments{
\item{response}{list, the standardised response}

\item{current_min_date}{datetime, the current min date obtained on the plan to detect if there are messages after this limit and discard them}
}
\description{
parses the response of bluesky API into the standardised episomer format. For more details see the vignette.
}
