% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R
\name{bluesky_rate_limited_check}
\alias{bluesky_rate_limited_check}
\title{function to check if there has been a rate limit error returned by the bluesky API}
\usage{
bluesky_rate_limited_check(resp)
}
\arguments{
\item{resp, }{the httr2 response object}
}
\description{
function to check if there has been a rate limit error returned by the bluesky API
}
