% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R
\name{bluesky_rerun_after_rate_limit}
\alias{bluesky_rerun_after_rate_limit}
\title{function to extract the number of seconds to wait after a rate limit has been reached}
\usage{
bluesky_rerun_after_rate_limit(resp)
}
\arguments{
\item{resp, }{the httr2 response object}
}
\description{
function to extract the number of seconds to wait after a rate limit has been reached
}
