% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_media.R
\name{create_api_and_plan_files_for_new_social_media}
\alias{create_api_and_plan_files_for_new_social_media}
\title{internal function for producing the code template for a new social media}
\usage{
create_api_and_plan_files_for_new_social_media(
  new_media,
  ref_media = "bluesky"
)
}
\arguments{
\item{new_media}{char, name of the social media to add}

\item{ref_media}{char, name of the social media to use as a reference to produce template}
}
\description{
internal function for producing the code template for a new social media
}
