% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-app.R
\name{dashboard_app}
\alias{dashboard_app}
\title{Run the episomer Shiny app}
\usage{
dashboard_app(data_dir = NA, host = NULL, port = NULL)
}
\arguments{
\item{data_dir}{Path to the 'data directory' containing application settings, models and collected tweets.
If not provided the system will try to reuse the existing one from last session call of \code{\link{setup_config}} or use the EPI_HOME environment variable, default: NA}

\item{host}{The host to run the Shiny app on, default: NULL (will run on 127.0.0.1)}

\item{port}{The port to run the Shiny app on, default: NULL (will run on a random port)}

\item{profile}{The profile to run the Shiny app on, default: "dashboard" (can be "dashboard" or "admin")}
}
\value{
The Shiny server object containing the launched application
}
\description{
Open the episomer Shiny app, used to setup the Data collection & processing pipeline, the Requirements & alerts pipeline and to visualise the outputs.
}
\details{
The episomer app is the user entry point to the episomer package. This application will help the user to setup the tweet collection process, manage all settings,
see the interactive dashboard visualisations, export them to Markdown or PDF, and setup the alert emails.

All its functionality is described on the episomer vignette.
}
\examples{
if(FALSE){
   #Running the episomer app
   library(episomer)
   message('Please choose the episomer data directory')
   setup_config(file.choose())
   episomer_app()
}
}
\seealso{
\code{\link{search_loop}}

 \code{\link{detect_loop}}
}
