% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_media.R
\name{is_r_folder_present}
\alias{is_r_folder_present}
\title{check whether the R folder is present in the current directory}
\usage{
is_r_folder_present()
}
\description{
check whether the R folder is present in the current directory
}
