% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{is_search_running}
\alias{is_search_running}
\title{Check whether the post collection task is running}
\usage{
is_search_running()
}
\value{
logical Whether the post collection is running
}
\description{
gets the post collection execution status
}
\details{
returns a logical value being TRUE if the post collection is running
}
\examples{
if(FALSE){
   library(episomer)
   message('Please choose the episomer data directory')
   setup_config(file.choose())
   is_search_running()
}
}
