% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_media.R
\name{keep_roxygen_and_function_declarations}
\alias{keep_roxygen_and_function_declarations}
\title{internal function for exporting code of necessary functions for creating templates for a new social media}
\usage{
keep_roxygen_and_function_declarations(file_content)
}
\arguments{
\item{file_content}{char, lines of the script to extract content from}
}
\description{
internal function for exporting code of necessary functions for creating templates for a new social media
}
