% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{missing_search_jobs}
\alias{missing_search_jobs}
\title{Missing search jobs if any}
\usage{
missing_search_jobs()
}
\value{
char list of social media with no search job running
}
\description{
check if search jobs for each activated social media are running and return the list names of social media with missing jobs
}
\details{
returns a char vector with the name of missing social media
}
\examples{
if(FALSE){
   library(episomer)
   message('Please choose the episomer data directory')
   setup_config(file.choose())
   missing_search_jobs()
}
}
