% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_topic}
\alias{search_topic}
\title{performs a search following the given plan, query and topic.}
\usage{
search_topic(plan, query, topic, sandboxed = FALSE)
}
\arguments{
\item{plan}{list, the plan that defines the timespan to search for}

\item{query}{char, the query to use when searching for the topic}

\item{topic}{char, the topic associated to this search}

\item{sandboxed}{logical, whether to un in sandboxed mode which prevent sending data to the embedded database.}
}
\description{
performs a search following the given plan, query and topic.
}
