% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_touch}
\alias{search_touch}
\title{updates the modificatin date of a flag file indicating that messages have been sent to the embedded database.}
\usage{
search_touch()
}
\description{
updates the modificatin date of a flag file indicating that messages have been sent to the embedded database.
}
