% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R, R/sm-bobol-api.R
\name{sm_api_get_token_bluesky}
\alias{sm_api_get_token_bluesky}
\alias{bluesky_create_session}
\alias{sm_api_get_token_bobol}
\title{Get token}
\usage{
sm_api_get_token_bluesky()

bluesky_create_session(handle = NULL, password = NULL)

sm_api_get_token_bobol()
}
\arguments{
\item{handle}{bluesky handle}

\item{password}{bluesky password}

\item{login_url}{bluesky login URL}
}
\value{
Token

List with access_jwt and did

Token
}
\description{
Get token

Get bearer token

Get token
}
\details{
You can find more information about the bluesky API here: \url{https://docs.bsky.app/docs/api/com-atproto-server-create-session}
}
\examples{
\dontrun{
bluesky_create_session()
}

}
