% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_bsky_auth}
\alias{set_bsky_auth}
\title{Save Blusky credentials (login and password) and store them securely}
\usage{
set_bsky_auth(bsky_user = "", bsky_password = "")
}
\arguments{
\item{bsky_user}{Username}

\item{bsky_password}{Password}
}
\description{
Update configuration object and underlying storage with given bluesky username and password. The password is encrypted.
}
\details{
Update Twitter authentication tokens in configuration object
}
\examples{
if(FALSE){
 #Setting the configuration values
   set_bsky_auth(
     bsky_user = "your user here",
     bsky_password = "your password",
   )
}
}
\seealso{
\code{\link{save_config}}
}
