% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_api_get_token}
\alias{sm_api_get_token}
\title{Get social media token}
\usage{
sm_api_get_token(network)
}
\arguments{
\item{network}{char, network the name of the social media}
}
\description{
Get social media token
}
\details{
this function implements the authethication on the respective social media and stores it securely. 
It is called by the search loop before starting message collection
}
