% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R
\name{sm_api_get_token_bluesky}
\alias{sm_api_get_token_bluesky}
\title{Get social media token}
\usage{
sm_api_get_token_bluesky()
}
\value{
Token
}
\description{
Get social media token
}
\details{
this function implements the authethication on the respective social media and stores it securely. 
It is called by the search loop before starting message collection
}
