% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R
\name{sm_api_got_rows_bluesky}
\alias{sm_api_got_rows_bluesky}
\title{Receive the results provided by a request and return a logical value indicating whether the request returned results.}
\usage{
sm_api_got_rows_bluesky(results)
}
\arguments{
\item{results}{list, results of the request in the standardised episomer format}
}
\description{
Receive the results provided by a request and return a logical value indicating whether the request returned results.
}
\details{
This function needs to be implemented for each social media since it is interpreted as 'No more messages are available for this plan'
When FALSE is returned the underlying plan will be ended
}
