% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_api_search}
\alias{sm_api_search}
\title{Search posts for a given query and produce standardised results}
\usage{
sm_api_search(query, token, plan)
}
\arguments{
\item{query}{char, the query to use in the search in nattive format}

\item{token}{char, the token to authenticate against the underlying social media api}

\item{plan}{list, the plan containing the targeted time frame}
}
\description{
Search posts for a given query and produce standardised results
}
\details{
This function impleents the search of messages for a given query and plan.
}
