% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_api_set_auth}
\alias{sm_api_set_auth}
\title{Save credentials for a social media provider and store them securely}
\usage{
sm_api_set_auth(network, shiny_input_list)
}
\arguments{
\item{network}{char, the name of the social media}

\item{shiny_input_list}{list, the shiny input object to extract the information from}
}
\description{
Update configuration object and underlying storage with given bluesky username and password. The password is encrypted.
}
