% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_api_translate_query}
\alias{sm_api_translate_query}
\title{Translate a parsed query to the social media query format}
\usage{
sm_api_translate_query(network, parsed)
}
\arguments{
\item{network}{char, the name of the social media}

\item{parsed}{char, parsed query to translate in one or many native equivalent queries}
}
\description{
Translate a parsed query to the social media query format
}
\details{
This function receive as a parameter a parsed query from the a topic in the list of topics. 
It can return one or many queries that matches the expected results using the underlying social media API
}
