% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_api_update_plan_after_request}
\alias{sm_api_update_plan_after_request}
\title{Takes a plan after a request has been performed and the standardised results provided as returned by the function sm_api_search_xxx
it should return a plan with necessary information updated so next request will continue paginating as expected to obtain
targeted messages}
\usage{
sm_api_update_plan_after_request(plan, results)
}
\arguments{
\item{plan}{list, the plan producing the resusts}

\item{results}{list, the result of the query in episomer format}
}
\description{
Takes a plan after a request has been performed and the standardised results provided as returned by the function sm_api_search_xxx
it should return a plan with necessary information updated so next request will continue paginating as expected to obtain
targeted messages
}
