% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R
\name{sm_api_update_plan_after_request_bluesky}
\alias{sm_api_update_plan_after_request_bluesky}
\title{Takes a plan after a request has been performed and the standardised results provided as returned by the function sm_api_search_xxx
it should return a plan with necessary information updated so next request will continue paginating as expected to obtain
targeted messages}
\usage{
sm_api_update_plan_after_request_bluesky(plan, results)
}
\arguments{
\item{plan}{list, the plan to update}

\item{results}{list, the obtained results of the plan}
}
\description{
Takes a plan after a request has been performed and the standardised results provided as returned by the function sm_api_search_xxx
it should return a plan with necessary information updated so next request will continue paginating as expected to obtain
targeted messages
}
