% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_plan_first_attributes}
\alias{sm_plan_first_attributes}
\title{returns a list with the custom social media plan attributes with the default values for a first plan in a topic.}
\usage{
sm_plan_first_attributes(network)
}
\arguments{
\item{network}{char, the name of the social media}
}
\description{
returns a list with the custom social media plan attributes with the default values for a first plan in a topic.
}
