% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-plan.R
\name{sm_plan_first_attributes_bluesky}
\alias{sm_plan_first_attributes_bluesky}
\title{returns a list with the custom social media plan attributes with the default values for a first plan in a topic.}
\usage{
sm_plan_first_attributes_bluesky()
}
\description{
returns a list with the custom social media plan attributes with the default values for a first plan in a topic.
}
