% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_plan_format}
\alias{sm_plan_format}
\title{prepare a plan for serialisation by transforming attributes with non standard serialisation types e.g. dates to string on the format expected by the parse functions}
\usage{
sm_plan_format(plan)
}
\arguments{
\item{plan, }{the plan to format the attributes}
}
\description{
prepare a plan for serialisation by transforming attributes with non standard serialisation types e.g. dates to string on the format expected by the parse functions
}
