% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-plan.R
\name{sm_plan_format_bluesky}
\alias{sm_plan_format_bluesky}
\title{prepare a plan for serialisation by transforming attributes with non standard serialisation types e.g. dates to string on the format expected by the parse functions}
\usage{
sm_plan_format_bluesky(plan)
}
\arguments{
\item{plan}{list, the plan with the attributes go format}
}
\description{
prepare a plan for serialisation by transforming attributes with non standard serialisation types e.g. dates to string on the format expected by the parse functions
}
