% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_plan_get_progress}
\alias{sm_plan_get_progress}
\title{calculates and returns the estimated progress of a plan using custom attributes}
\usage{
sm_plan_get_progress(plan)
}
\arguments{
\item{plan}{list, the plan to obtain the progress from}
}
\description{
calculates and returns the estimated progress of a plan using custom attributes
}
