% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-plan.R
\name{sm_plan_get_progress_bluesky}
\alias{sm_plan_get_progress_bluesky}
\title{calculates and returns the estimated progress of a plan using custom attributes}
\usage{
sm_plan_get_progress_bluesky(plan)
}
\arguments{
\item{plan}{list, the plan to calculate the progress from}
}
\description{
calculates and returns the estimated progress of a plan using custom attributes
}
