% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_plan_next_attributes}
\alias{sm_plan_next_attributes}
\title{returns a list with the custom social media plan attributes with the values for the plan to be run after the plans provided as parameter}
\usage{
sm_plan_next_attributes(network, plans)
}
\arguments{
\item{network}{char, network the name of the social media}

\item{plans}{list, list of plans to use for calculate new plan attributes}
}
\description{
returns a list with the custom social media plan attributes with the values for the plan to be run after the plans provided as parameter
}
