% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-plan.R
\name{sm_plan_next_attributes_bluesky}
\alias{sm_plan_next_attributes_bluesky}
\title{returns a list with the custom social media plan attributes with the values for the plan to be run after the plans provided as parameter}
\usage{
sm_plan_next_attributes_bluesky(plans)
}
\arguments{
\item{plans}{list, the list with current plans of the given topic to calculate next plan attributes}
}
\description{
returns a list with the custom social media plan attributes with the values for the plan to be run after the plans provided as parameter
}
