% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_plan_next_search_info}
\alias{sm_plan_next_search_info}
\title{Return a message to log what the next query is going to be for the current plan}
\usage{
sm_plan_next_search_info(plan)
}
\arguments{
\item{plan}{list, the plan to obtain the message}
}
\description{
Return a message to log what the next query is going to be for the current plan
}
