% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm.R
\name{sm_plan_parse_attributes}
\alias{sm_plan_parse_attributes}
\title{parse custom attributes for a plan in this specific social media and returns them as a named list with attributes in the proper type}
\usage{
sm_plan_parse_attributes(network, ...)
}
\arguments{
\item{network}{char, the name of the social media}

\item{...}{list, triple dot arguments to be passed as a parameter tu underlying functions}
}
\description{
parse custom attributes for a plan in this specific social media and returns them as a named list with attributes in the proper type
}
