% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-plan.R
\name{sm_plan_parse_attributes_bluesky}
\alias{sm_plan_parse_attributes_bluesky}
\title{parse custom attributes for a plan in this specific social media and returns them as a named list with attributes in the proper type}
\usage{
sm_plan_parse_attributes_bluesky(...)
}
\arguments{
\item{...}{list, triple dot arguments to be passed as a parameter tu underlying functions}
}
\description{
parse custom attributes for a plan in this specific social media and returns them as a named list with attributes in the proper type
}
