% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-plan.R
\name{sm_plan_search_info_bluesky}
\alias{sm_plan_search_info_bluesky}
\title{Return a message to log what the next query is going to be for the current plan}
\usage{
sm_plan_search_info_bluesky(plan)
}
\arguments{
\item{plan}{list, the plan to print the information about}
}
\description{
Return a message to log what the next query is going to be for the current plan
}
