% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduler.R
\name{stop_search_runner_task}
\alias{stop_search_runner_task}
\title{Stops the post collection task}
\usage{
stop_search_runner_task()
}
\value{
Nothing
}
\description{
stops the post collection task
}
\details{
Stops the post collection task if it is already running
This function will try also deactivate the respective scheduled task on Windows.
}
\examples{
if(FALSE){
   #getting tasks statuses
   library(episomer)
   message('Please choose the episomer data directory')
   setup_config(file.choose())
   stop_search_runner_task()
}
}
