% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_media.R
\name{add_new_social_media}
\alias{add_new_social_media}
\title{Add new social media files}
\usage{
add_new_social_media(social_media, social_media_ref = "bluesky")
}
\arguments{
\item{social_media}{Name of the new social media}

\item{social_media_ref}{Name of the reference social media}
}
\description{
Add new social media files
}
