% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R, R/sm-bobol-api.R
\name{sm_api_set_auth_bluesky}
\alias{sm_api_set_auth_bluesky}
\alias{sm_api_set_auth_bobol}
\title{Save bluesky credentials (login and password) and store them securely}
\usage{
sm_api_set_auth_bluesky(shiny_input_list)

sm_api_set_auth_bobol(shiny_input_list)
}
\arguments{
\item{shiny_input_list}{List of shiny input values containing the bobol specific credentials}
}
\description{
Update configuration object and underlying storage with given bluesky username and password. The password is encrypted.

Update configuration object and underlying storage with given bobol username and password. The password is encrypted.
}
\details{
Update authentication tokens in configuration object

Update authentication tokens in configuration object
}
\examples{
if(FALSE){
 #Setting the configuration values
   sm_api_set_auth_bluesky(
     shiny_input_list = list(bluesky_user = "your user here", bluesky_password = "your password"),
   )
}
if(FALSE){
 #Setting the configuration values
   sm_api_set_auth_bobol(
     shiny_input_list = list(bobol_user = "your user here", bobol_password = "your password"),
   )
}
}
\seealso{
\code{\link{save_config}}

\code{\link{save_config}}
}
