% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm-bluesky-api.R
\name{sm_api_search_bluesky}
\alias{sm_api_search_bluesky}
\title{Search posts for a given query}
\usage{
sm_api_search_bluesky(query, token, plan, max_retries = 20, verbose = TRUE)
}
\arguments{
\item{query}{query to search for}

\item{token}{Access token}

\item{plan}{Plan object}

\item{max_retries}{Maximum number of retries}

\item{verbose}{Whether to print progress messages}
}
\value{
List with posts and next cursor for resumption
}
\description{
Search posts for a given query
}
\details{
You can find more information about the bluesky API here: \url{https://docs.bsky.app/docs/api/app-bsky-feed-search-posts}
Additionnaly more information about the "post" object is available here: \url{https://atproto.blue/en/latest/atproto/atproto_client.models.app.bsky.feed.defs.html#atproto_client.models.app.bsky.feed.defs.PostView}
}
